refactor : GameResult의 생성 책임을 Ladder 에서 제거
